Create CodeWars - Find The Odd Int
My solution to the problem, originally used an ArrayList which could work if multiple odd numbers were required. However as only one number was used an odd amount of times in the examples the solution was ok.
The class commented out below was a really interesting solution I found and need to look more into.