Create CodeWars - Multi-tap Keypad Text Entry on an Old Mobile Phone
Tried two approaches to solving this, one checking two strings that registered as the 3 letter keypads or 4 letter keypads and then tried getting the index of it and the modulus of that index to find the number it was at but was unsuccessful. Possibly could have done that with more thought into it. 

The commented out class below is one of the better solutions, should consider using something like that for future reference :)